use linptech_receiver instead R3AC and R4-X
