+++ README.md
@@ -3,8 +3,7 @@
 ## 1.支持硬件
 
 - 领普产品
-  - [x] R3AC+天线，可以90m以上
-  - [x] 四路接收器+外置天线
+  - [x] 接收器：R3AC+四路接收器
   - [ ] K4R实现总控区域接收器，一个对多个接收器，距离品质稳定性有待提高
   - [ ] K5单开或G1S发射器，桌面一对一或者多对一控制一个接收器，基本没有问题
   - [ ] 中继器，中继控制命令（来自发射器或者控制中心）到接收器，中继状态值到控制中心
